Fit growth model to bipod object

Fit growth model to bipod object

Usage

fit(
  x,
  growth_type = "exponential",
  infer_t0 = TRUE,
  variational = FALSE,
  factor_size = 1,
  model_selection_algo = "bayes_factor",
  chains = 4,
  iter = 5000,
  cores = 4
)

Arguments

x

a bipod object

growth_type

character string specifying the type of growth assumed, one of 'exponential', 'logistic', or 'both'. If 'both' is selected model selection will be performed.

infer_t0

Boolean, if TRUE an estimate of the instant of time in which the population is born will be inferred

variational

Boolean specifying whether using variational as opposed to mcmc sampling

factor_size

numeric factor by which to divide counts in the bipod object

model_selection_algo

Algorithm to use for model selection, either 'bayes_factor' or 'mixture_model'

chains

integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm

iter

integer number of iterations to run in the MCMC algorithm

cores

integer number of cores to use in parallel processing

Value

the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit