Fit breakpoints to bipod object

Fit breakpoints to bipod object

Usage

fit_breakpoints(
  x,
  norm = TRUE,
  n_trials = 500,
  avg_points_per_window = 3,
  available_changepoints = c(0:5),
  model_selection = "LOO",
  n_core = 4
)

Arguments

x

A bipod object.

norm

Logical value indicating whether to normalize the data. If TRUE, the time and count data are standardized before fitting the model. (default is TRUE)

n_trials

Integer specifying the number of trials for the optimization algorithm. This controls the number of iterations used to fit the breakpoints. (default is 500)

avg_points_per_window

Integer specifying the average number of data points per segment. This parameter influences the granularity of the segments when searching for breakpoints. (default is 3)

available_changepoints

Integer vector specifying the range of available changepoints. These values represent the possible number of breakpoints to be considered during model fitting. (default is 0:5)

model_selection

Character string specifying the model selection criterion. Options include 'LOO' (Leave-One-Out cross-validation), 'AIC' (Akaike Information Criterion), and 'BIC' (Bayesian Information Criterion). (default is "LOO")

n_core

Integer specifying the number of CPU cores to use for parallel processing. (default is 4)

Value

The input bipod object with an added 'breakpoints_fit' slot containing the fitted model for the breakpoints.