Plot Omega Values from Mixture Model Fit

Generates a plot of the omega values obtained from fitting a mixture model to the data in a bipod object. The plot type can be customized to show a histogram, violin plot, or boxplot.

Usage

plot_mixture_model_omega(x, plot_type = "hist", color = "maroon")

Arguments

x

A bipod object. Must contain a 'fit' field and the 'omega_mixture_model' field in its metadata.

plot_type

A character string specifying the type of plot to generate. Options are "hist" for histogram, "violin" for violin plot, and "boxplot" for boxplot. (default is 'hist')

color

A character string specifying the color to use for the plot elements. (default is 'maroon')

Value

A ggplot2 object displaying the plot of omega values.