Plot traces of specified parameters from a fitted Stan model

Plot traces of specified parameters from a fitted Stan model

Usage

plot_traces(x, fit, pars = c(), diagnose = FALSE)

Arguments

x

A biPOD object of class bipod. Must contains at least one fit

fit

the name of the fit oh which the traces should be plotted

pars

A character vector of parameters to plot.

diagnose

A Boolean indicating whether the plots should be colored and contain info regarding the convergence of the MCMC sampling.

Value

A ggplot object with traces of the specified parameters